+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v1.1.1
 
-- [x] fix validation issues
+- [x] fix bug when css nesting is disabled #89
+- [x] validation rules update #87
 
 ## v1.1.0